+++ examples/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 # [1.2.0-alpha.14](https://github.com/didi/LogicFlow/compare/examples@1.2.0-alpha.13...examples@1.2.0-alpha.14) (2022-07-13)
+## [1.1.21](https://github.com/didi/LogicFlow/compare/examples@1.1.20...examples@1.1.21) (2022-07-02)
+
+**Note:** Version bump only for package examples
+
+
+
+
+
+## [1.1.20](https://github.com/didi/LogicFlow/compare/examples@1.1.19...examples@1.1.20) (2022-06-08)
 
 **Note:** Version bump only for package examples